requirements.txt
gradio
pandas
plotly
requests
python-dotenv
numpy
python-multipart